feat: optimise bundle size
