+++ packages/sdk-ts/src/core/accounts/index.ts
@@ -2,5 +2,3 @@ export * from './PrivateKey'
 export * from './PublicKey'
 export * from './Address'
 export * from './BaseAccount'
-export * from './signers'
-export { accountParser as injectiveAccountParser } from './AccountParser'
