+++ packages/sdk-ts/src/core/index.ts
@@ -1,4 +1,3 @@
 export * from './modules'
 export * from './accounts'
 export * from './utils'
-export * as InjectiveStargate from './stargate'
