+++ packages/sdk-ts/src/core/modules/ibc/index.ts
@@ -1,4 +1,3 @@
 import MsgTransfer from './msgs/MsgTransfer'
-import MsgTransferCosmjs from './msgs/MsgTransferCosmjs'
 
-export { MsgTransfer, MsgTransferCosmjs }
+export { MsgTransfer }
