+++ packages/sdk-ts/src/exports.ts
@@ -1,0 +1,6 @@
+import MsgTransferCosmjs from './core/modules/ibc/msgs/MsgTransferCosmjs'
+import { accountParser as injectiveAccountParser } from './core/accounts/AccountParser'
+
+export * from './core/accounts/signers'
+export * from './core/stargate'
+export { MsgTransferCosmjs, injectiveAccountParser }
