+++ packages/ts-types/package.json
@@ -4,11 +4,12 @@
   "version": "1.14.11-beta.2",
   "sideEffects": false,
   "license": "Apache-2.0",
+  "types": "dist/cjs/index.d.ts",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
   "author": {
-    "name": "Bojan Angjelkoski",
-    "email": "bojan@injectivelabs.org"
+    "name": "InjectiveLabs",
+    "email": "admin@injectivelabs.org"
   },
   "files": [
     "dist"
