+++ packages/utils/package.json
@@ -4,10 +4,11 @@
   "version": "1.14.11-beta.6",
   "sideEffects": false,
   "author": {
-    "name": "Bojan Angjelkoski",
-    "email": "bojan@injectivelabs.org"
+    "name": "InjectiveLabs",
+    "email": "admin@injectivelabs.org"
   },
   "license": "Apache-2.0",
+  "types": "dist/cjs/index.d.ts",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
   "files": [
