+++ packages/wallet-ts/src/exports.ts
@@ -1,0 +1,1 @@
+export * as UtilsWallets from './utils/wallets'
