+++ packages/wallet-ts/src/strategies/wallet-strategy/strategies/Keplr.ts
@@ -10,7 +10,6 @@ import {
   TxResponse,
   waitTxBroadcasted,
   createTxRawFromSigResponse,
-  createCosmosSignDocFromSignDoc,
   createSignDocFromTransaction,
 } from '@injectivelabs/sdk-ts'
 import type { DirectSignResponse } from '@cosmjs/proto-signing'
@@ -29,6 +28,7 @@ import {
   WalletEventListener,
 } from '../../../types/enums'
 import { SendTransactionOptions } from '../types'
+import { createCosmosSignDocFromSignDoc } from '../../../utils/cosmos'
 
 export default class Keplr extends BaseConcreteStrategy implements ConcreteWalletStrategy {
   private keplrWallet: KeplrWallet
