+++ packages/wallet-ts/src/strategies/wallet-strategy/strategies/Ninji.ts
@@ -16,7 +16,6 @@ import {
   TxResponse,
   waitTxBroadcasted,
   createTxRawFromSigResponse,
-  createCosmosSignDocFromSignDoc,
   createSignDocFromTransaction,
 } from '@injectivelabs/sdk-ts'
 import type { DirectSignResponse } from '@cosmjs/proto-signing'
@@ -29,6 +28,7 @@ import {
   WalletEventListener,
 } from '../../../types/enums'
 import { SendTransactionOptions } from '../types'
+import { createCosmosSignDocFromSignDoc } from '../../../utils/cosmos'
 
 export default class Ninji extends BaseConcreteStrategy implements ConcreteWalletStrategy {
   private ninjiWallet: NinjiWallet
