+++ packages/wallet-ts/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './enums'
+
 export type UnwrappedPromise<T> = T extends Promise<infer Return> ? Return : T
