+++ packages/wallet-ts/src/utils/index.ts
@@ -1,3 +1,2 @@
-export * as UtilsWallets from './wallets'
 export * from './wallets/cosmos'
 export * from './utils'
